+++ lib/util/extensions/npc/npc_extensions.dart
@@ -84,9 +84,7 @@ extension NpcExtensions on Npc {
         observed?.call();
         moveTowardsTarget(
           target: player,
-          close: () {
-            closePlayer(player);
-          },
+          close: () => closePlayer(player),
           margin: margin,
         );
       },